MotherSon
Problem Statement:

Develop a personal portfolio website 

Requirements:
A fully functional personal portfolio website as per the Figma design.
The website should have a responsive design and be optimized for different devices.

Steps to Complete the Task:

Understanding Design Requirements:
Access and review the provided Figma design to gain a clear understanding of the project requirements.

Navigate to Projects:
When you click on the "Project" button, the page should navigate to the project's section. 

Adding Projects:
There is a section to fill in project details. After entering the details, clicking the "Add" button should add the project to the "Projects View" section of the webpage.

Menu Navigation:
The webpage has several buttons: About, Projects, and Contact.
Clicking on the "About" button should take you to the About section of the page.
Clicking on the "Projects" button should take you to the Projects section.
Clicking on the "Contact" button should take you to the Contact section.

LinkedIn Button:
There is also a LinkedIn button, but for now, it's just for display and doesn't perform any function.

Test and refine:
Conduct thorough testing to ensure that the project-adding functionality works as intended.

Finalize and submit:
Clean up the code, refactor if necessary, and optimize performance.
Figma:
https://www.figma.com/file/PXaJ8K2iI5cdGQNY7VjRT8/Personal-Portfolio-Website?type=design&node-id=0%3A1&mode=design&t=ovk3f2TuCvgltBs9-1

Submission Instructions:

Use HTML, CSS, JS, and React to complete the assignment.
Publish/Deploy the task and provide us the published URL.
Make a 5-minute video with self introduction and explaining the features you added to the assignment. You can use Loom to record it or record directly with a phone camera and upload it to Google Drive. Share the video link with public access.
Submit the assignment & video using the below form: